GH-2291 Added changeThreshold test
