+++ MultisigIntegrationTests/SafeTransactionControllerIntegrationTests.swift
@@ -64,7 +64,7 @@ class SafeTransactionControllerIntegrationTests: CoreDataTestCase {
 
         let exp = expectation(description: "proposing")
 
-        SafeTransactionController.shared.proposeTransaction(
+        proposeTransaction(
                 transaction: tx,
                 sender: proposingOwner,
                 chainId: chainId,
@@ -116,7 +116,7 @@ class SafeTransactionControllerIntegrationTests: CoreDataTestCase {
 
         let exp = expectation(description: "proposing")
 
-        SafeTransactionController.shared.proposeTransaction(
+        proposeTransaction(
                 transaction: tx,
                 sender: proposingOwner,
                 chainId: chainId,
@@ -141,38 +141,15 @@ class SafeTransactionControllerIntegrationTests: CoreDataTestCase {
         let privateKey = try PrivateKey(data: Data(hex: "0xe7979e5f2ceb1d4ef76019d1fdba88b50ceefe0575bbfdf94969837c50a5d895"))
         let proposingOwner: Address = "0x728cafe9fB8CC2218Fb12a9A2D9335193caa07e0"
         let safeAddress: Address = "0x193BF1F9655eAA511d2255c0efF1D7693c59d77F"
-        //let threshold = try Sol.UInt256(UInt256(1).data32) // 1 should not fail
-        let threshold = try Sol.UInt256(UInt256(0).data32) // 0 should fail when executed
-        //let threshold = try Sol.UInt256(UInt256(20).data32) // 20 should fail when executed
-
-//        let changeThresholdABI = GnosisSafe_v1_3_0.changeThreshold(
-//                _threshold: threshold
-//        ).encode()
-//
-//        let tx = Transaction(safeAddress: safeAddress,
-//                chainId: chainId,
-//                toAddress: safeAddress,
-//                contractVersion: "1.3.0",
-//                amount: "0",
-//                data: changeThresholdABI,
-//                safeTxGas: "0",
-//                nonce: "0")
-
-//        guard let tx = tx else {
-//            XCTFail("TX not created")
-//            return
-//        }
-
-//        let signature = try SafeTransactionSigner().sign(tx, key: privateKey)
-//        let transactionSignature = signature.hexadecimal
+        let threshold = 1
 
         let exp = expectation(description: "proposing") //?
-        let safe = Safe.by(address: safeAddress.checksummed, chainId: "4")
-        let tx = SafeTransactionController.shared.changeThreshold(safe: safe, safeTxGas: "0", nonce: "0", threshold: threshold)
+        let safe = createSafe(name: "foo", address: safeAddress.checksummed, chain: try! makeChain(id: "4"))
+        let tx = SafeTransactionController.shared.changeThreshold(safe: safe, safeTxGas: "0", nonce: "0", threshold: threshold)!
         let signature = try SafeTransactionSigner().sign(tx, key: privateKey)
         let transactionSignature = signature.hexadecimal
 
-        SafeTransactionController.shared.proposeTransaction(
+        proposeTransaction(
                 transaction: tx,
                 sender: proposingOwner,
                 chainId: chainId,
@@ -210,4 +187,31 @@ class SafeTransactionControllerIntegrationTests: CoreDataTestCase {
         waitForExpectations(timeout: 15)
     }
 
+    // Helper methods
+    func proposeTransaction(transaction: Transaction,
+                            sender: Address,
+                            chainId: String,
+                            signature: String,
+                            completion: @escaping (Result<SCGModels.TransactionDetails, Error>) -> Void) -> URLSessionTask? {
+        let task = App.shared.clientGatewayService.asyncProposeTransaction(transaction: transaction,
+                sender: AddressString(sender),
+                signature: signature,
+                chainId: chainId) { result in
+            DispatchQueue.main.asyncAfter(deadline: .now() + .milliseconds(600)) {
+                switch result {
+                case .failure(let error):
+                    if (error as NSError).code == URLError.cancelled.rawValue &&
+                               (error as NSError).domain == NSURLErrorDomain {
+                        // Estimation was canceled, ignore.
+                        return
+                    }
+
+                default: break
+                }
+
+                completion(result)
+            }
+        }
+        return task
+    }
 }
